Remove unnecessary AbstractBlockStore
